refactor: address various PR review comments
see https://github.com/opentripplanner/otp-react-redux/pull/359
