+++ lib/components/narrative/narrative-itineraries.js
@@ -1,10 +1,8 @@
 import coreUtils from '@opentripplanner/core-utils'
-import { getCompanyIcon } from '@opentripplanner/icons/lib/companies'
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import Skeleton, { SkeletonTheme } from 'react-loading-skeleton'
 import { connect } from 'react-redux'
-import styled from 'styled-components'
 
 import {
   setActiveItinerary,
@@ -15,50 +13,17 @@ import {
   updateItineraryFilter
 } from '../../actions/narrative'
 import * as uiActions from '../../actions/ui'
-import Icon from '../narrative/icon'
+import NarrativeItinerariesErrors from './narrative-itineraries-errors'
+import NarrativeItinerariesHeader from './narrative-itineraries-header'
 import { ComponentContext } from '../../util/contexts'
 import {
   getActiveItineraries,
   getActiveSearch,
-  getErrorMessage,
   getRealtimeEffects,
   getResponsesWithErrors
 } from '../../util/state'
 
-import SaveTripButton from './save-trip-button'
-
 const { ItineraryView } = uiActions
-
-const IssueContainer = styled.div`
-  border-top: 1px solid grey;
-  display: flex;
-  padding: 10px;
-`
-
-const IssueIconContainer = styled.div`
-  font-size: 18px;
-  height: 24px;
-  width: 24px;
-`
-
-const IssueContents = styled.div`
-  font-size: 14px;
-  margin-left: 10px;
-  text-align: left;
-`
-
-const IssueButton = styled.button`
-  background-color: #ECBE03;
-  border: none;
-  border-radius: 5px;
-  display: inline-block;
-  font-size: 12px;
-  padding: 4px;
-
-  span {
-    margin-left: 4px;
-  }
-`
 
 class NarrativeItineraries extends Component {
   static propTypes = {
@@ -113,144 +78,16 @@ class NarrativeItineraries extends Component {
     updateItineraryFilter({sort: {...sort, direction}})
   }
 
-  _toggleRealtimeItineraryClick = (e) => {
-    const {setUseRealtimeResponse, useRealtime} = this.props
-    setUseRealtimeResponse({useRealtime: !useRealtime})
+  _onViewAllOptions = () => {
+    if (this.props.itineraryIsExpanded) {
+      this._toggleDetailedItinerary()
+    } else {
+      this._toggleShowErrors()
+    }
   }
 
   _toggleShowErrors = () => {
     this.setState({ showingErrors: !this.state.showingErrors })
-  }
-
-  _renderErrors = () => {
-    const { errorMessages, errors } = this.props
-    return errors.map((error, idx) => {
-      let icon
-      if (error.network) {
-        const CompanyIcon = getCompanyIcon(error.network)
-        icon = <CompanyIcon />
-      } else {
-        icon = <Icon className='text-warning' type='exclamation-triangle' />
-      }
-      return (
-        <IssueContainer key={idx}>
-          <IssueIconContainer>
-            {icon}
-          </IssueIconContainer>
-          <IssueContents>
-            {getErrorMessage(error, errorMessages)}
-          </IssueContents>
-        </IssueContainer>
-      )
-    })
-  }
-
-  _renderErrorsButton = () => {
-    const { errors } = this.props
-    if (errors.length === 0) return null
-    return (
-      <IssueButton onClick={this._toggleShowErrors}>
-        <Icon type='warning' />
-        <span>{errors.length} issues</span>
-      </IssueButton>
-    )
-  }
-
-  _renderHeader = () => {
-    const {
-      errors,
-      itineraries,
-      itineraryIsExpanded,
-      pending,
-      sort
-    } = this.props
-    const { showingErrors } = this.state
-
-    let resultText, titleText
-    if (pending) {
-      resultText = 'Finding your options...'
-      titleText = 'Finding your options...'
-    } else {
-      const itineraryPlural = itineraries.length === 1
-        ? 'itinerary'
-        : 'itineraries'
-      const issuePlural = errors.length === 1
-        ? 'issue'
-        : 'issues'
-      resultText = `${itineraries.length} ${itineraryPlural} found.`
-      titleText = errors.length > 0
-        ? `${itineraries.length} ${itineraryPlural} (and ${errors.length} ${issuePlural}) found`
-        : resultText
-    }
-
-    let content
-
-    if (itineraryIsExpanded || showingErrors) {
-      content = (
-        <>
-          <button
-            className='clear-button-formatting'
-            onClick={itineraryIsExpanded
-              ? this._toggleDetailedItinerary
-              : this._toggleShowErrors
-            }
-          >
-            <Icon type='arrow-left' /> View all options
-          </button>
-
-          {itineraryIsExpanded && <SaveTripButton />}
-        </>
-      )
-    } else {
-      content = (
-        <>
-          <div
-            style={{flexGrow: 1}}
-            title={titleText}
-          >
-            <span style={{
-              marginRight: '10px'
-            }}>
-              {resultText}
-            </span>
-            {this._renderErrorsButton()}
-          </div>
-          <div style={{display: 'flex', float: 'right'}}>
-            <button
-              className='clear-button-formatting'
-              onClick={this._onSortDirChange}
-              style={{marginRight: '5px'}}
-            >
-              <Icon type={`sort-amount-${sort.direction.toLowerCase()}`} />
-            </button>
-            <select
-              onBlur={this._onSortChange}
-              onChange={this._onSortChange}
-              value={sort.value}
-            >
-              <option value='BEST'>Best option</option>
-              <option value='DURATION'>Duration</option>
-              <option value='ARRIVALTIME'>Arrival time</option>
-              <option value='DEPARTURETIME'>Departure time</option>
-              <option value='WALKTIME'>Walk time</option>
-              <option value='COST'>Cost</option>
-            </select>
-          </div>
-        </>
-      )
-    }
-
-    return (
-      <div
-        className='options header'
-        style={{
-          alignItems: 'end',
-          display: 'flex'
-        }}
-      >
-        {content}
-      </div>
-    )
   }
 
   _renderLoadingDivs = () => {
@@ -279,8 +116,11 @@ class NarrativeItineraries extends Component {
       activeSearch,
       activeStep,
       containerStyle,
+      errorMessages,
+      errors,
       itineraries,
       itineraryIsExpanded,
+      pending,
       realtimeEffects,
       setActiveItinerary,
       setActiveStep,
@@ -302,48 +142,20 @@ class NarrativeItineraries extends Component {
       !useRealtime
     )
 
-    let content
-
-    if (showingErrors) {
-      content = this._renderErrors()
-    } else {
-      content = (
-        <>
-          {itineraries.map((itinerary, index) => {
-            const active = index === activeItinerary
-            // Hide non-active itineraries.
-            if (!active && itineraryIsExpanded) return null
-            return (
-              <ItineraryBody
-                active={active}
-                activeLeg={activeLeg}
-                activeStep={activeStep}
-                expanded={showDetails}
-                index={index}
-                itinerary={itinerary}
-                key={index}
-                LegIcon={LegIcon}
-                onClick={active ? this._toggleDetailedItinerary : undefined}
-                routingType='ITINERARY'
-                setActiveItinerary={setActiveItinerary}
-                setActiveLeg={this._setActiveLeg}
-                setActiveStep={setActiveStep}
-                setVisibleItinerary={setVisibleItinerary}
-                showRealtimeAnnotation={showRealtimeAnnotation}
-                sort={sort}
-                timeFormat={timeFormat}
-                visibleItinerary={visibleItinerary}
-              />
-            )
-          })}
-          {this._renderLoadingDivs()}
-        </>
-      )
-    }
-
     return (
       <div className='options itinerary' style={containerStyle}>
-        {this._renderHeader()}
+        <NarrativeItinerariesHeader
+          errors={errors}
+          itineraries={itineraries}
+          itineraryIsExpanded={itineraryIsExpanded}
+          onSortChange={this._onSortChange}
+          onSortDirChange={this._onSortDirChange}
+          onToggleShowErrors={this._toggleShowErrors}
+          onViewAllOptions={this._onViewAllOptions}
+          pending={pending}
+          showingErrors={showingErrors}
+          sort={sort}
+        />
         <div
           // FIXME: Change to a ul with li children?
           className='list'
@@ -352,7 +164,44 @@ class NarrativeItineraries extends Component {
             overflowY: 'auto'
           }}
         >
-          {content}
+          {showingErrors
+            ? (
+              <NarrativeItinerariesErrors
+                errorMessages={errorMessages}
+                errors={errors}
+              />
+            )
+            : <>
+              {itineraries.map((itinerary, index) => {
+                const active = index === activeItinerary
+                // Hide non-active itineraries.
+                if (!active && itineraryIsExpanded) return null
+                return (
+                  <ItineraryBody
+                    active={active}
+                    activeLeg={activeLeg}
+                    activeStep={activeStep}
+                    expanded={showDetails}
+                    index={index}
+                    itinerary={itinerary}
+                    key={index}
+                    LegIcon={LegIcon}
+                    onClick={active ? this._toggleDetailedItinerary : undefined}
+                    routingType='ITINERARY'
+                    setActiveItinerary={setActiveItinerary}
+                    setActiveLeg={this._setActiveLeg}
+                    setActiveStep={setActiveStep}
+                    setVisibleItinerary={setVisibleItinerary}
+                    showRealtimeAnnotation={showRealtimeAnnotation}
+                    sort={sort}
+                    timeFormat={timeFormat}
+                    visibleItinerary={visibleItinerary}
+                  />
+                )
+              })}
+              {this._renderLoadingDivs()}
+            </>
+          }
         </div>
       </div>
     )
