+++ lib/util/state.js
@@ -69,9 +69,7 @@ function humanReadableMode (modeStr) {
 }
 
 /**
- * Generates a list of issues with a list of OTP responses. For batch routing
- * this could be an array of multiple responses, but for standard itinerary
- * routing it will be at most an array with one response.
+ * Generates a list of issues/errors from a list of OTP responses.
  *
  * @param {Object} otpState the OTP state object
  * @return {Array} An array of objects describing the errors seen. These will at
@@ -122,11 +120,12 @@ export function getResponsesWithErrors (otpState) {
 }
 
 /**
- * Gets the active error (first OTP plan response that contains an error). This
+ * Gets the active error (eiter a planning error or a rental feed error). This
  * should only be used with itinerary routing, not batch routing.
  */
 export function getActiveError (otpState) {
-  return getResponsesWithErrors(otpState)[0]
+  const errors = getResponsesWithErrors(otpState)
+  return errors[errors.length - 1]
 }
 
 /**
